+++ CHANGES.rst
@@ -6,6 +6,10 @@ Version 3.1.0
 Unreleased
 
 -  Support Cookie CHIPS (Partitioned Cookies). :issue:`2797`
+-   ``CacheControl.no_transform`` is a boolean when present. ``min_fresh`` is
+    ``None`` when not present. Added the ``must_understand`` attribute. Fixed
+    some typing issues on cache control. :issue:`2881`
+
 
 Version 3.0.3
 -------------